+++ libhdf5/hdf5dispatch.c
@@ -103,6 +103,7 @@ static const NC_Dispatch HDF5_dispatcher = {
     NC4_HDF5_set_var_chunk_cache,
     NC4_get_var_chunk_cache,
 
+    NC4_filter_actions,    
 };
 
 const NC_Dispatch* HDF5_dispatch_table = NULL; /* moved here from ddispatch.c */